requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
Pillow